code cleanup on user management
